created a separate Results component and ResultsStep that is linked to vibe check steps
